pyproject.toml
[tool.poetry]
name = "LIN-IN"
version = "0.0.1"
description = "LinkedIn Insightes using a data export"
authors = [
    "toni <toni.fetzer@cronn.de>"
    ]
license = "GPL-3"

readme = "README.md"
repository = "https://github.com/python-poetry/poetry"

[tool.poetry.dependencies]
python = "3.12.1"
networkx = "^2.8.6"
matplotlib = "^3.5.3"
nltk = "^3.7"
numpy = "^1.23.3"
pandas = "^1.4.4"
Pillow = "^10.0.0"
plotly = "^5.10.0"
pyjanitor = "^0.23.1"
pyvis = "^0.2.1"
streamlit = "^1.12.2"
thefuzz = "^0.19.0"
wordcloud = "^1.8.2"
python-Levenshtein = "^0.12.2"

[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
